moving yql queries to separate class
