+++ calculate/badboystats.py
@@ -13,7 +13,7 @@ class BadBoyStats(object):
 
         self.rankings = {}
         # Load the scoring based on crime categories
-        with open("crimecatsscoring.csv", mode="r", encoding="utf-8-sig") as infile:
+        with open("resources/crimecatsscoring.csv", mode="r", encoding="utf-8-sig") as infile:
             reader = csv.reader(infile)
             for rows in reader:
                 cat = rows[0].upper().strip()
